requirements.txt
textual
langchain==1.2.8
langchain-openai
langchain-mcp-adapters
langgraph-checkpoint-sqlite
python-dotenv
langchain-community
duckduckgo-search
ddgs
faiss-cpu
sentence-transformers
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-go
tree-sitter-rust
pathspec
watchdog
black
isort
pytest
pytest-cov
pytest-asyncio